Revert "Merge branch 'feature/refactoring' into develop"
This reverts commit c15aa682477a061ce009403851e1a57699fe210b, reversing
changes made to 3ec168009a765dc8c3dca9aa5277585c588c2b71.
