--- frontend/src/components/Input/InputsDefine.jsx
@@ -1,70 +1,0 @@
-export const codeInput = {
-  id: "code",
-  name: "code",
-  type: "text",
-  placeholder: "인증 코드",
-};
-
-export const emailInput = {
-  id: "email",
-  name: "email",
-  type: "email",
-  placeholder: "Email",
-};
-
-export const nameInput = {
-  id: "name",
-  name: "name",
-  type: "text",
-  placeholder: "Name",
-};
-
-export const passwordInput = {
-  id: "password",
-  name: "password",
-  type: "password",
-  placeholder: "Password",
-};
-
-export const secondPasswordInput = {
-  id: "secondPassword",
-  name: "secondPassword",
-  type: "password",
-  placeholder: "PassCheck Passwordword",
-};
-export const femaleInput = {
-  id: "female",
-  name: "gender",
-  type: "radio",
-  value: "FEMALE",
-};
-
-export const maleInput = {
-  id: "male",
-  name: "gender",
-  type: "radio",
-  value: "MALE",
-};
-
-export const birthdayInput = { id: "birthday", name: "birthday", type: "date" };
-
-export const securityEditInput = [
-  {
-    id: "pastPassword",
-    name: "pastPassword",
-    type: "password",
-    placeholder: "pastPassword",
-  },
-  {
-    id: "password",
-    name: "password",
-    type: "password",
-    placeholder: "Password",
-  },
-  {
-    id: "secondPassword",
-    name: "secondPassword",
-    type: "password",
-    placeholder: "Check Password",
-  },
-];
