+++ frontend/src/components/Inputs.jsx
@@ -1,0 +1,70 @@
+export const registerInput = [
+  {
+    id: "name",
+    name: "name",
+    type: "text",
+    placeholder: "Name",
+  },
+  {
+    id: "password",
+    name: "password",
+    type: "password",
+    placeholder: "Password",
+  },
+  {
+    id: "secondPassword",
+    name: "secondPassword",
+    type: "password",
+    placeholder: "Check Password",
+  },
+];
+export const genderInput = [
+  {
+    id: "female",
+    name: "gender",
+    type: "radio",
+    value: "FEMALE",
+  },
+  {
+    id: "male",
+    name: "gender",
+    type: "radio",
+    value: "MALE",
+  },
+];
+export const loginInput = [
+  {
+    id: "email",
+    name: "email",
+    type: "email",
+    placeholder: "Email",
+  },
+
+  {
+    id: "password",
+    name: "password",
+    type: "password",
+    placeholder: "Password",
+  },
+];
+
+export const securityEditInput = [
+  {
+    id: "pastPassword",
+    name: "pastPassword",
+    type: "password",
+    placeholder: "pastPassword",
+  },
+  {
+    id: "password",
+    name: "password",
+    type: "password",
+    placeholder: "Password",
+  },
+  {
+    id: "secondPassword",
+    name: "secondPassword",
+    type: "password",
+    placeholder: "Check Password",
+  },
+];
