+++ frontend/src/components/Schema.jsx
