+++ frontend/src/screens/LoginScreen.jsx
@@ -1,13 +1,10 @@
 import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import LoginForm from "../components/LoginForm";
+import { useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { loginAction } from "../actions/userAction";
-import { LoginFormHook } from "../hooks/FormHoooks";
-import { loginSchema } from "../components/Form/Schema";
 
 const LoginSceen = () => {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
 
   /** Get loginInfo from Redux(loginInfo를 redux에서 가져옴)*/
   const loginInfo = useSelector((state) => state.loginInfo);
@@ -19,17 +16,6 @@ const LoginSceen = () => {
     }
   }, [loginInfo?.sessfbs_ffa0934]);
 
-  /** Getting input data via Submit(Submit을 통해 input data를 가져옴)*/
-  const onSubmit = (data) => {
-    /** Send data to loginAction after getting the params(params를 받고 난 후 loginAction으로 보냄)*/
-    dispatch(
-      loginAction({
-        mail: data.email,
-        password: data.password,
-      })
-    );
-  };
-
   return (
     <section className="form">
       <div className="form__container">
@@ -40,11 +26,7 @@ const LoginSceen = () => {
               Broaden your connection via VISTA
             </span>
           </div>
-          <LoginFormHook
-            schema={loginSchema}
-            onSubmit={onSubmit}
-            loginInfo={loginInfo}
-          />
+          <LoginForm loginInfo={loginInfo} />
         </div>
       </div>
